move 'anyPass' tests to separate file
