+++ types/ramda/tsconfig.json
@@ -27,6 +27,7 @@
         "test/allPass-tests.ts",
         "test/always-tests.ts",
         "test/and-tests.ts",
-        "test/any-tests.ts"
+        "test/any-tests.ts",
+        "test/anyPass-tests.ts"
     ]
 }
